prdigitos-filipe/requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
pillow==9.5.0
opencv-python==4.8.0.76
numpy==1.26.0
streamlit-drawable-canvas==0.9.3
plotly==5.15.0